+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # V0.4 01.05.2024
 - multivendor implementation. Now different vendors can login and use their IDs. The IDs and PW are saved in users.json
 - landscape mode crashed on later android versions. No error to be seen. Maybe an emulator problem.
+- Need to get categories of menu. I need to get only the first category, as it is an array. The print should be ordered by categories
 
 # V0.2 15.03.2024
 UI updated, Popup for print reset added.